backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-docx==1.2.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
greenlet==3.1.1
anthropic>=0.49.0
bcrypt>=4.0.0
PyJWT[crypto]==2.9.0
aiosqlite>=0.20.0
slowapi==0.1.9
openpyxl==3.1.5
rank-bm25==0.2.2
snowballstemmer==3.0.1
PyYAML==6.0.3
jinja2==3.1.6
